Linux compilation fix

+++ Samples/Compositor/src/Compositor.cpp
@@ -268,7 +268,7 @@ void Sample_Compositor::checkBoxToggled(OgreBites::CheckBox * box)
 					for (size_t i=0; i<numTextures; i++)
 					{
 						//Dirty string composition. NOT ROBUST!
-						mDebugTextureSelectMenu->addItem(compositorName + ";" + texDef->name + ";" + i);
+						mDebugTextureSelectMenu->addItem(compositorName + ";" + texDef->name + ";" + (uint32)i);
 					}
 				}
 				else
